+++ app/src/main/AndroidManifest.xml
@@ -9,13 +9,13 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".SuddenPlusInput">
+        <activity android:name=".activities.SuddenPlusInputActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".SuddenPlusTable" android:parentActivityName=".SuddenPlusInput" />
+        <activity android:name=".activities.SuddenPlusTableActivity" android:parentActivityName=".activities.SuddenPlusInputActivity" />
     </application>
 
 </manifest>
