+++ app/src/main/java/com/duraz0rz/suddenpluscalculator/views/SuddenPlusTableView.kt
@@ -1,0 +1,6 @@
+package com.duraz0rz.suddenpluscalculator.views
+
+class SuddenPlusTableView {
+
+
+}
