+++ app/src/main/res/layout/activity_sudden_plus_table.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.duraz0rz.suddenpluscalculator.SuddenPlusTable">
+    tools:context="com.duraz0rz.suddenpluscalculator.activities.SuddenPlusTableActivity">
 
     <TableLayout
         android:layout_width="match_parent"
