+++ app/src/test/java/com/duraz0rz/suddenpluscalculator/views/SuddenPlusTableViewTest.kt
@@ -1,0 +1,4 @@
+package com.duraz0rz.suddenpluscalculator.views
+
+class SuddenPlusTableViewTest {
+}
